requirements.txt
flask==2.0.1
pillow==9.0.0
numpy==1.22.0
gunicorn==20.1.0
mysql-connector-python==8.0.28
python-dotenv==0.19.2